adding a world full of colors
the colour hue (in the HSV color model) is used to colour bugs
according to priority, using the standard colours of the heat spectrum
(e.g. P1=red, P2=orange, P3=yellow, P4=green, P5=blue)
the colour saturation and/or value (in the HSV color model) is used
to indicate the age of bugs

So an old P1 would be a faded red, but a new P4 would be bright green.

Closes #11