Show the same data as before on the html views
Due to the changes done to the methods to save PR
the index and repo views were not showing the data
to accompany their titles. This fixes it for the
moment by restricting the data provided to only
open PRs. This should be fixed in the future as
some of the endpoints mention open or all but they
are not really showing open or all only.