Moved: creation/update of the products table
The config defines the "actual state", which includes the products to
fetch and display. Changing the config shouldn't require a
nailed --migrate, since the database layout didn't change. Because of
this, I moved the code to create/update the products table to the
beginning of the nailed script, so that a simple restart of nailed would
be enough.