Changed: SQL statement to show open bugs per day
Before the change, nailed showed all bugs weirdly aggregated to 20 data
points for all the available data, which resulted in the loss of
information of the number of open bugs currently.

Now nailed will show all bugs open for each day, where the latest data
point collected on that day is chosen to represent the whole day. Thus,
the latest data point in the graph is also the most current data point
available.