Modify app.rb to support gitlab
app.rb now supports the use of multiple version
control systems.
Adapt views/layout.haml to the changes.
Note: This also fixes a bug with missing routes for
repositories if there no database entry for that repo.
Repositories will only be displayed in the header if there
is actual data about them in the database.

Signed-off-by: Pascal Arlt <parlt@suse.com>