Add Gitlab to index page's table
Add `Merge Requests` column to the index page's table.
The column contains a list of open gitlab merge requests
similar to the one for github PR's.
Just like for the header the column will only appear if
there are open changerequests.
Add a mouseover indicator when hovering the column names
to clarify wether the column corresponds to github or
gitlab.

Signed-off-by: Pascal Arlt <parlt@suse.com>